Inline and break out more server code
